Better update messaging
When not showing the UI, fuel rods will only be updated when a variance
of more than 50 fuel units (2% of an ingot) appears. This is about 30
seconds of normal operation and significantly reduces network traffic
without compromising look&feel.

Also fixing heat so that heat is radiated along the entire length of the
fuel column. This is more expensive, but necessary for complex reactor
interiors.