Issue #40: Blutonium could only be added by hand
Basically fixing the UI slots to use the same system as the Inventories,
as there was a data but - I'd registered graphite bars as the equivalent
of blutonium fuel, but the UI worked via the oredict.

Fixing the registration fixes the issue, also changing how slots work to
be more robust in the future, so I won't have to fiddle with slots when
I add more fuels/wastes.