fixed possible crash
i once got a ClassCastException crash caused by casting a TileEntityReactorFuelRod to a TileEntityReactorControlRod. i havent experimented but i assume it happens if someone places a fuel rod where a control rod is meant to be. but that wouldnt make much sense because ive placed all sorts of block in there by mistake sometimes. but checking if the tileentity is a control rod should hopefully fix the rare crash.
the crash report if you want to read: https://pastebin.com/D81r33Fy